pyproject.toml
[project]
name = "google-mcp-server"
version = "0.1.0"
description = "Google Suite MCP Server"
authors = [
    {name = "Aarya Ranjit",email = "aaryaranjit03@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.11.0)",
    "attrs (==25.4.0)",
    "cachetools (==6.2.1)",
    "certifi (==2025.10.5)",
    "cffi (==2.0.0)",
    "charset-normalizer (==3.4.4)",
    "click (==8.3.0)",
    "colorama (==0.4.6)",
    "cryptography (==46.0.3)",
    "future (==1.0.0)",
    "google-api-core (==2.28.1)",
    "google-api-python-client (==2.187.0)",
    "google-auth (==2.41.1)",
    "google-auth-httplib2 (==0.2.1)",
    "google-auth-oauthlib (==1.2.3)",
    "googleapis-common-protos (==1.72.0)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httplib2 (==0.31.0)",
    "httpx (==0.28.1)",
    "httpx-sse (==0.4.3)",
    "idna (==3.11)",
    "jsonschema (==4.25.1)",
    "jsonschema-specifications (==2025.9.1)",
    "lxml (==5.1.0)",
    "mcp (==1.21.0)",
    "mpmath (==1.3.0)",
    "numpy (==2.1.2)",
    "oauthlib (==3.3.1)",
    "packaging (==24.1)",
    "pillow (==10.4.0)",
    "proto-plus (==1.26.1)",
    "protobuf (==6.33.0)",
    "psutil (==6.0.0)",
    "py-cpuinfo (==9.0.0)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pycparser (==2.23)",
    "pydantic (==2.12.4)",
    "pydantic-settings (==2.12.0)",
    "pydantic-core (==2.41.5)",
    "pyjwt (==2.10.1)",
    "pyngrok (==7.3.0)",
    "pyparsing (==3.2.0)",
    "pyqt5-qt5 (==5.15.2)",
    "pyqt5-sip (==12.13.0)",
    "pyserial (==3.5)",
    "python-dotenv (==1.2.1)",
    "python-multipart (==0.0.20)",
    "pytz (==2024.1)",
    "pywin32 (==311)",
    "pyyaml (==6.0.1)",
    "referencing (==0.37.0)",
    "requests (==2.32.5)",
    "requests-oauthlib (==2.0.0)",
    "rpds-py (==0.28.0)",
    "setuptools (==75.1.0)",
    "six (==1.16.0)",
    "sniffio (==1.3.1)",
    "sse-starlette (==3.0.3)",
    "sympy (==1.13.3)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "tzdata (==2024.2)",
    "uritemplate (==4.2.0)",
    "urllib3 (==2.2.3)",
    "uvicorn (==0.38.0)",
    "watchdog (==6.0.0)",
    "fastapi (>=0.121.1,<0.122.0)",
    "redis (>=7.0.1,<8.0.0)",
    "google-genai (>=1.50.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
